+++ scene/register_scene_types.cpp
@@ -173,6 +173,13 @@
 #include "scene/resources/skeleton_modification_2d_twoboneik.h"
 #include "scene/resources/skeleton_modification_stack_2d.h"
 #include "scene/resources/skeleton_modification_3d.h"
+#include "scene/resources/skeleton_modification_3d_ccdik.h"
+#include "scene/resources/skeleton_modification_3d_fabrik.h"
+#include "scene/resources/skeleton_modification_3d_jiggle.h"
+#include "scene/resources/skeleton_modification_3d_lookat.h"
+#include "scene/resources/skeleton_modification_3d_stackholder.h"
+#include "scene/resources/skeleton_modification_3d_twoboneik.h"
+#include "scene/resources/skeleton_modification_stack_3d.h"
 #include "scene/resources/sky.h"
 #include "scene/resources/sky_material.h"
 #include "scene/resources/sphere_shape_3d.h"
