+++ scene/resources/skeleton_modification_stack_3d.h
@@ -1,5 +1,5 @@
 /*************************************************************************/
-/*  skeleton_modification_3d.h                                           */
+/*  skeleton_modification_stack_3d.h                                     */
 /*************************************************************************/
 /*                       This file is part of:                           */
 /*                           GODOT ENGINE                                */
@@ -28,49 +28,63 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                */
 /*************************************************************************/
 
-#ifndef SKELETONMODIFICATION3D_H
-#define SKELETONMODIFICATION3D_H
+#ifndef SKELETONMODIFICATIONSTACK3D_H
+#define SKELETONMODIFICATIONSTACK3D_H
 
 #include "scene/3d/skeleton_3d.h"
-#include "scene/resources/skeleton_modification_stack_3d.h"
 
-class SkeletonModificationStack3D;
+class Skeleton3D;
+class SkeletonModification3D;
 
-class SkeletonModification3D : public Resource {
-	GDCLASS(SkeletonModification3D, Resource);
+class SkeletonModificationStack3D : public Resource {
+	GDCLASS(SkeletonModificationStack3D, Resource);
 	friend class Skeleton3D;
-	friend class SkeletonModificationStack3D;
+	friend class SkeletonModification3D;
 
 protected:
 	static void _bind_methods();
-
-	SkeletonModificationStack3D *stack;
-	int execution_mode = 0; // 0 = process
-
-	bool enabled = true;
-	bool is_setup = false;
-	bool execution_error_found = false;
-
-	bool _print_execution_error(bool p_condition, String p_message);
+	void _get_property_list(List<PropertyInfo> *p_list) const;
+	bool _set(const StringName &p_path, const Variant &p_value);
+	bool _get(const StringName &p_path, Variant &r_ret) const;
 
 public:
-	virtual void execute(float delta);
-	virtual void setup_modification(SkeletonModificationStack3D *p_stack);
+	Skeleton3D *skeleton;
+	bool is_setup = false;
+	bool enabled = true;
+	float strength = 1.0;
 
-	float clamp_angle(float angle, float min_bound, float max_bound, bool invert);
+	enum EXECUTION_MODE {
+		execution_mode_process,
+		execution_mode_physics_process,
+	};
 
-	void set_enabled(bool p_enabled);
-	bool get_enabled();
+	Vector<Ref<SkeletonModification3D>> modifications;
+	int modifications_count = 0;
 
-	void set_execution_mode(int p_mode);
-	int get_execution_mode() const;
+	void setup();
+	void execute(float delta, int execution_mode);
 
-	Ref<SkeletonModificationStack3D> get_modification_stack();
+	void enable_all_modifications(bool p_enable);
+	Ref<SkeletonModification3D> get_modification(int p_mod_idx) const;
+	void add_modification(Ref<SkeletonModification3D> p_mod);
+	void delete_modification(int p_mod_idx);
+	void set_modification(int p_mod_idx, Ref<SkeletonModification3D> p_mod);
 
-	void set_is_setup(bool p_setup);
+	void set_modification_count(int p_count);
+	int get_modification_count() const;
+
+	void set_skeleton(Skeleton3D *p_skeleton);
+	Skeleton3D *get_skeleton() const;
+
 	bool get_is_setup() const;
 
-	SkeletonModification3D();
+	void set_enabled(bool p_enabled);
+	bool get_enabled() const;
+
+	void set_strength(float p_strength);
+	float get_strength() const;
+
+	SkeletonModificationStack3D();
 };
 
-#endif // SKELETONMODIFICATION3D_H
+#endif // SKELETONMODIFICATIONSTACK3D_H
